Methods list

Sound.Sound

Sound.pick_output

Sound.record

Sound.sound

Sound.sound

Sound.sound

Sound.sound

Sound.sound

Sound.sound

Sound.soundsc

Sound.soundsc

Methods usage

Sound.Sound — Module

Sound

Module that exports wrappers around audio methods in PortAudio.

sound and soundsc : methods for audio playback.

record : method for audio recording.

Sound.pick_output — Method

pick_output( ; io_in::IO = stdin, io_out::IO = stdout)

Show available audio devices and prompt user to select output device.

Sound.record — Function

data, S = record(time::Real = 5; args=(1,0), chat::Bool=true)

Record time seconds of audio data using the built-in microphone.

Input

time : duration; 5 seconds by default

Option

args : arguments to PortAudioStream; (1,0) for single-channel input by default

chat : show begin/end message? true by default

Output

data : Vector of length time * S

S : sample_rate of stream

Sound.sound — Function

sound(x::AbstractVector, S::Real = framerate(x), args...; kwargs...)

Play monophonic audio signal x at sampling rate S samples per second through default audio output device using the PortAudio package. Caller must specify S unless a framerate method is defined for x.

Sound.sound — Function

sound(output_device_index::Int, x, S::Real = framerate(x); kwargs...)

Play sound in x using devices()[output_device_index].

Sound.sound — Function

sound(:pick, x, S::Real = framerate(x); kwargs...)

Prompt user to pick output device, then play sound in x.

Sound.sound — Function

sound(x::AbstractMatrix, S::Real = framerate(x) [, output_device])

Play stereo audio signal x at sampling rate S samples per second through default audio output device using the PortAudio package. Caller must specify S unless a framerate method is defined for x.

Sound.sound — Method

sound(sb:SampleBuf)

Play audio signal sb of type SampleBuf through default audio output device using the PortAudio package.

Sound.sound — Method

sound(:first))

Return first audio device that has output channels.

Sound.soundsc — Function

soundsc(x, S::Real = framerate(x), args...; kwargs...)

Call sound after scaling x to have values in (-1,1).

Sound.soundsc — Method

soundsc(sb:SampleBuf)

Play audio signal sb of type SampleBuf through default audio output device using the PortAudio package, after scaling the data to have values in (-1,1).