Methods list

Sound.sound

Sound.sound

Sound.sound

Methods usage

Sound.sound — Method

sound(x)

Play mono or stereo audio signal x at default sampling rate 8192 samples per second through default audio output device using the PortAudio package.

Sound.sound — Method

sound(x::AbstractMatrix, S::Real)

Play stereo audiosignal x at sampling rate S samples per second through default audio output device using the PortAudio package.

Sound.sound — Method

sound(x::AbstractVector, S::Real)

Play monophonic audiosignal x at sampling rate S samples per second through default audio output device using the PortAudio package.